split out version.go from app.go
